Change the exam card background to be a FrameLayout with an ImageView and a View
Previously, the card background was an ImageView with an image as a background ('android:background') and a white gradient as its source
('android:src'). This caused the image to stretch even though the 'android:scaleType' property was set to 'centerCrop'. The reason this happened is because the 'android:scaleType' property only effects the 'android:src', which in our case was the white gradient.

Now, the background image and the white gradient are separated, so the 'android:scaleType' now effects the image and not the gradient.